refactor(install): expose pure manifest planner
